Move InitialFileManager into internal package
