+++ bundles/org.eclipse.cdt.lsp/OSGI-INF/org.eclipse.cdt.lsp.internal.editor.InitialFileManager.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="org.eclipse.cdt.lsp.internal.editor.InitialFileManager">
+   <service>
+      <provide interface="org.eclipse.cdt.lsp.editor.InitialUri"/>
+   </service>
+   <reference cardinality="1..1" field="workspace" interface="org.eclipse.core.resources.IWorkspace" name="workspace"/>
+   <implementation class="org.eclipse.cdt.lsp.internal.editor.InitialFileManager"/>
+</scr:component>
