+++ bundles/org.eclipse.cdt.lsp/src/org/eclipse/cdt/lsp/internal/editor/InitialFileManager.java
@@ -11,13 +11,15 @@
  * Alexander Fedorov (ArSysOp) - rework to OSGi component
  *******************************************************************************/
 
-package org.eclipse.cdt.lsp.plugin;
+package org.eclipse.cdt.lsp.internal.editor;
 
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Optional;
 
 import org.eclipse.cdt.lsp.ExistingResource;
+import org.eclipse.cdt.lsp.editor.InitialUri;
+import org.eclipse.cdt.lsp.plugin.LspPlugin;
 import org.eclipse.core.resources.IWorkspace;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.Platform;
